requirements.txt
Django==5.1.4
pillow==11.1.0
django-cleanup==9.0.0
django-allauth==65.3.1
django-htmx==1.21.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
uvicorn==0.34.0
python-dotenv==1.0.1
django-admin-honeypot-2024==1.0.0
whitenoise==6.8.2
django-cloudinary-storage==0.3.0
django-resized==1.0.3
celery==5.4.0
redis==2.0.1
flower==5.2.1
gevent==24.11.1
django-celery-beat==2.7.0